Removed explicit casting by the QGauss instances' constructors.

+++ source/navier_stokes_projection/assemble_advection_matrix.cc
@@ -14,7 +14,7 @@ void NavierStokesProjection<dim>::assemble_velocity_advection_matrix()
   // Polynomial degree of the integrand
   const int p_degree = 3 * velocity.fe_degree - 1;
 
-  const QGauss<dim>   quadrature_formula(std::ceil(0.5 * (p_degree + 1)));
+  const QGauss<dim>   quadrature_formula(std::ceil(0.5 * double(p_degree + 1)));
 
   using CellFilter =
     FilteredIterator<typename DoFHandler<dim>::active_cell_iterator>;
