+++ source/navier_stokes_projection/assemble_mass_and_laplace_matrices.cc
@@ -16,7 +16,7 @@ void NavierStokesProjection<dim>::assemble_velocity_matrices()
   // Polynomial degree of the integrand
   const int p_degree = 2 * velocity.fe_degree;
 
-  const QGauss<dim>   quadrature_formula(std::ceil(0.5 * (p_degree + 1)));
+  const QGauss<dim>   quadrature_formula(std::ceil(0.5 * double(p_degree + 1)));
 
   using CellFilter =
     FilteredIterator<typename DoFHandler<dim>::active_cell_iterator>;
@@ -138,7 +138,7 @@ void NavierStokesProjection<dim>::assemble_pressure_matrices()
   // Polynomial degree of the integrand
   const int p_degree = 2 * pressure.fe_degree;
 
-  const QGauss<dim>   quadrature_formula(std::ceil(0.5 * (p_degree + 1)));
+  const QGauss<dim>   quadrature_formula(std::ceil(0.5 * double(p_degree + 1)));
 
   using CellFilter =
     FilteredIterator<typename DoFHandler<dim>::active_cell_iterator>;
