+++ source/navier_stokes_projection/assemble_poisson_prestep_rhs.cc
@@ -19,14 +19,14 @@ assemble_poisson_prestep_rhs()
 
   const int p_degree = p_degree_body_force + pressure.fe_degree - 1;
 
-  const QGauss<dim>   quadrature_formula(std::ceil(0.5 * (p_degree + 1)));
+  const QGauss<dim>   quadrature_formula(std::ceil(0.5 * double(p_degree + 1)));
 
   // Polynomial degree of the boundary integrand
 
   const int face_p_degree = std::max(pressure.fe_degree + p_degree_body_force,
                                      pressure.fe_degree + velocity.fe_degree -2);
 
-  const QGauss<dim-1>   face_quadrature_formula(std::ceil(0.5 * (face_p_degree + 1)));
+  const QGauss<dim-1>   face_quadrature_formula(std::ceil(0.5 * double(face_p_degree + 1)));
 
   using CellFilter =
     FilteredIterator<typename DoFHandler<dim>::active_cell_iterator>;
